requirements.txt
Flask
Flask_SQLAlchemy
Pillow
opencv-python
matplotlib
python-dotenv 